sb: fix k_predict not building under msvc

+++ softbody/s_compute_cuda_kernels.cu
@@ -338,7 +338,7 @@ __global__ void k_predict(
     // Velocity damping
     float3 v0_predamp = xyz(d_velocities[id]);
     float3 ang_v0_predamp = xyz(d_angular_velocities[id]);
-    float d = 1 / pow(2, dt);
+    float d = 1 / powf(2, dt);
     float3 v0 = d * v0_predamp;
     float3 ang_v0 = d * ang_v0_predamp;
